chore: forge fmt

+++ contracts/src/StabilityPool.sol
@@ -305,7 +305,7 @@ contract StabilityPool is LiquityBase, Ownable, CheckContract, IStabilityPool {
         uint256 currentETHGain = getDepositorETHGain(msg.sender);
         uint256 currentYieldGain = getDepositorYieldGain(msg.sender);
         uint256 compoundedBoldDeposit = getCompoundedBoldDeposit(msg.sender);
-        
+
         uint256 boldLoss = initialDeposit - compoundedBoldDeposit; // Needed only for event log
         (uint256 keptYieldGain, uint256 yieldGainToSend) = _getYieldToKeepOrSend(currentYieldGain, _doClaim);
 
@@ -316,7 +316,7 @@ contract StabilityPool is LiquityBase, Ownable, CheckContract, IStabilityPool {
 
         _depositBoldtoSP(msg.sender, _topUp);
         _decreaseYieldGainsOwed(currentYieldGain);
-        
+
         _sendBoldtoDepositor(msg.sender, yieldGainToSend);
         _stashOrSendETHGains(msg.sender, currentETHGain, boldLoss, _doClaim);
 
@@ -327,11 +327,11 @@ contract StabilityPool is LiquityBase, Ownable, CheckContract, IStabilityPool {
     function _getYieldToKeepOrSend(uint256 _currentYieldGain, bool _doClaim) internal pure returns (uint256, uint256) {
         uint256 yieldToKeep;
         uint256 yieldToSend;
-         
+
         if (_doClaim) {
             yieldToKeep = 0;
             yieldToSend = _currentYieldGain;
-        } else { 
+        } else {
             yieldToKeep = _currentYieldGain;
             yieldToSend = 0;
         }
@@ -352,7 +352,7 @@ contract StabilityPool is LiquityBase, Ownable, CheckContract, IStabilityPool {
         _requireUserHasDeposit(initialDeposit);
 
         activePool.mintAggInterest();
-     
+
         uint256 currentETHGain = getDepositorETHGain(msg.sender);
         uint256 currentYieldGain = getDepositorYieldGain(msg.sender);
 
@@ -407,11 +407,11 @@ contract StabilityPool is LiquityBase, Ownable, CheckContract, IStabilityPool {
         _requireUserHasNoDeposit(msg.sender);
         assert(getDepositorETHGain(msg.sender) == 0);
         assert(getDepositorYieldGain(msg.sender) == 0);
-       
+
         activePool.mintAggInterest();
 
         uint256 ETHToSend = _getTotalETHGainAndZeroStash(msg.sender, 0);
-        
+
         _sendETHGainToDepositor(ETHToSend);
 
         assert(stashedETH[msg.sender] == 0);
@@ -423,7 +423,7 @@ contract StabilityPool is LiquityBase, Ownable, CheckContract, IStabilityPool {
 
     function triggerBoldRewards(uint256 _boldYield) external {
         _requireCallerIsActivePool();
-   
+
         uint256 totalBoldDepositsCached = totalBoldDeposits; // cached to save an SLOAD
         /*
         * When total deposits is 0, B is not updated. In this case, the BOLD issued can not be obtained by later
@@ -431,11 +431,12 @@ contract StabilityPool is LiquityBase, Ownable, CheckContract, IStabilityPool {
         *
         */
         if (totalBoldDepositsCached == 0 || _boldYield == 0) {
-            return;}
+            return;
+        }
 
         yieldGainsOwed += _boldYield;
 
-        uint256 yieldPerUnitStaked =_computeYieldPerUnitStaked(_boldYield, totalBoldDepositsCached);
+        uint256 yieldPerUnitStaked = _computeYieldPerUnitStaked(_boldYield, totalBoldDepositsCached);
 
         uint256 marginalYieldGain = yieldPerUnitStaked * P;
         epochToScaleToB[currentEpoch][currentScale] = epochToScaleToB[currentEpoch][currentScale] + marginalYieldGain;
@@ -443,7 +444,7 @@ contract StabilityPool is LiquityBase, Ownable, CheckContract, IStabilityPool {
         emit B_Updated(epochToScaleToB[currentEpoch][currentScale], currentEpoch, currentScale);
     }
 
-    function _computeYieldPerUnitStaked(uint256 _yield, uint256 _totalBoldDeposits) internal returns (uint) {
+    function _computeYieldPerUnitStaked(uint256 _yield, uint256 _totalBoldDeposits) internal returns (uint256) {
         /*  
         * Calculate the BOLD-per-unit staked.  Division uses a "feedback" error correction, to keep the 
         * cumulative error low in the running total B:
@@ -599,7 +600,7 @@ contract StabilityPool is LiquityBase, Ownable, CheckContract, IStabilityPool {
         if (_amount == 0) return;
         uint256 newYieldGainsOwed = yieldGainsOwed - _amount;
         yieldGainsOwed = newYieldGainsOwed;
-        emit YieldGainsOwedUpdated(newYieldGainsOwed); 
+        emit YieldGainsOwedUpdated(newYieldGainsOwed);
     }
 
     // --- Reward calculator functions for depositor ---
